Adds alarm tableview to main screen

+++ Surf.Alarm/UI/SurfAlarmBuilder/SurfAlarmBuilderVC.swift
@@ -96,7 +96,7 @@ extension SurfAlarmBuilderVC: UITableViewDataSource, UITableViewDelegate {
 }
 
 extension SurfAlarmBuilderVC: SurfHeightSliderDelegate, SurfAlarmDaySelectionDelegate {
-    func minimumHeightSelectionChanged(to newHeight: Double) {
+    func minimumHeightSelectionChanged(to newHeight: Int) {
         self.alarm.minHeight = newHeight
     }
