+++ Surf.Alarm/UI/SurfAlarmBuilder/SurfHeightSelectorCell.swift
@@ -14,16 +14,16 @@ class SurfHeightSelectorCell: UITableViewCell {
     
     @IBAction func heightChanged(_ sender: UISlider) {
         if sender.value == 0 {
-            self.delegate?.minimumHeightSelectionChanged(to: 0.0)
+            self.delegate?.minimumHeightSelectionChanged(to: 0)
             heightLabel.text = "No Min"
         } else {
             let ftValue = Int((sender.value * 10).rounded())
-            self.delegate?.minimumHeightSelectionChanged(to: Double(ftValue))
+            self.delegate?.minimumHeightSelectionChanged(to: ftValue)
             heightLabel.text = "\(ftValue)+ ft"
         }
     }
 }
 
 protocol SurfHeightSliderDelegate: class {
-    func minimumHeightSelectionChanged(to newHeight: Double)
+    func minimumHeightSelectionChanged(to newHeight: Int)
 }
